pyproject.toml
[project]
name = "discsync-embeddings"
version = "1.0.19"
description = "Vectorize your favorite chats."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "PyYAML>=6.0",
    "httpx>=0.27.2",
    "psutil>=7.0.0",
    "boto3>=1.34.0",
    "PyGithub>=2.4.0",
    "python-dotenv>=1.0.1",
    "SQLAlchemy>=2.0.35",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "sqlmodel>=0.0.16",
    "greenlet>=3.2.4",
    "kubernetes>=34.1.0",
    "alembic>=1.13.2",
    "psycopg[binary,pool]>=3.2.11",
    "qdrant-client>=1.9.2",
    "sentence-transformers>=2.7.0",
    "numpy>=1.26.0",
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "llama-index-vector-stores-qdrant>=0.2.0",
]


[dependency-groups]
test = [
    "pytest",
]
dev = [
    "ruff",
    "hatchling",
]

[tool.uv]
package = true

[tool.uv.sources]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
testpaths = [
    "tests"
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["discsync_embeddings"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]
